backend_django/requirements.txt
# Django backend requirements
django>=5.0,<6.0
djangorestframework>=3.14,<4.0
pymongo>=4.6,<5.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0,<2.0
openpyxl>=3.1,<4.0
PyJWT>=2.8,<3.0
bleach>=6.1,<7.0